// react/src/core/publizon/model/libraryProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Publizon Library API
 * Pubhub exists in two separate environments, each with their own server, code and database. Please use the web service located at library-api.qa.pubhub.dk when developing and testing.
 * Orders/loans created in test environment will not be invoiced. Please request a new license key for the production environment when you're ready to go live.
 * OpenAPI spec version: 1
 */
import type { TestCard } from "./testCard";

export interface LibraryProfile {
  id?: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  isilNumber?: string | null;
  ebookLoanDurationInDays?: number;
  maxEbookAmountPerMonth?: number;
  /** @nullable */
  maxEbookAmountPerMonthNotificationThreshold?: number | null;
  maxConcurrentEbookLoansPerBorrower?: number;
  maxConcurrentEbookReservationsPerBorrower?: number;
  maxEbookCancellationsPerMonth?: number;
  audioLoanDurationInDays?: number;
  maxAudioAmountPerMonth?: number;
  /** @nullable */
  maxAudioAmountPerMonthNotificationThreshold?: number | null;
  maxConcurrentAudioLoansPerBorrower?: number;
  maxConcurrentAudioReservationsPerBorrower?: number;
  maxAudioCancellationsPerMonth?: number;
  /** @nullable */
  testCards?: TestCard[] | null;
}